Improve the handling of the SWI name translation for AArch64 APIs.
The SWI name translation to a C function name was close but not quite
right. It's now really really close. There's a couple of instances that
don't work exactly the same, but they're unlikely to affect anyone.
So this new translated form should work correctly for all the things
we are going to use it for.