Add creation of non-x-variant SWI interfaces.
All the SWIs now should be able to have both X-variant and non-X
variant SWIs defined, with the correct returns being given in each
case. I'm not completely sure that this is working, but the output
code looks correct. It's possible that we can optimise the outputs
as well, but again, that's not been checked.