Improve performance and fix for bad return values.
We no longer use the stack to store the intermediate return values,
and instead move into the other registers in the locally preserved
set (x8-x15). We use x17 for the error return, as that's free right
now, and we're more careful about marking what's happening with other
registers.

There were some problems with the x_variant settings being inverted.
These should now be fixed and I think we're now doing about as
efficient a job as we can in the veneer. Which, honestly, is a little
bit wasteful of my time as the system calls themselves always take
way longer than the setup. But it does mean that the single stepping
is a lot simpler for me.